+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * **[Feature]** Expose the ID Token and Access Token (as raw JWT format) in the `UserInformation` object returned from the sign-in method.
 * **[Fix]** Fix missing proguard rules so that the app does not have to specify them.
 * **[Fix]** Fix crash on silently refreshing token if initialization of MSAL fails.
+* **[Breaking change]** The `UserInformation` class has been moved from the `appcenter` module to the `appcenter-auth` module and must now be imported as `import com.microsoft.appcenter.auth.UserInformation`.
 
 ### App Center Data